environmentd: allow testing by other crates
This allows the balancer crate to spin up envd and mock frontegg servers
for testing.
