+++ src/environmentd/tests/sql.rs
@@ -93,6 +93,7 @@ use chrono::{DateTime, Utc};
 use http::StatusCode;
 use itertools::Itertools;
 use mz_adapter::{TimestampContext, TimestampExplanation};
+use mz_environmentd::test_util::{self, MzTimestamp, PostgresErrorExt, Server, KAFKA_ADDRS};
 use mz_ore::assert_contains;
 use mz_ore::now::{NowFn, NOW_ZERO, SYSTEM_TIME};
 use mz_ore::result::ResultExt;
@@ -109,10 +110,6 @@ use timely::order::PartialOrder;
 use tokio::sync::{mpsc, oneshot};
 use tokio_postgres::error::{DbError, SqlState};
 use tracing::{debug, info};
-
-use crate::util::{MzTimestamp, PostgresErrorExt, Server, KAFKA_ADDRS};
-
-pub mod util;
 
 /// An HTTP server whose responses can be controlled from another thread.
 struct MockHttpServer {
@@ -171,7 +168,7 @@ fn test_no_block() {
     // (as opposed to a CI timeout) causes `services.log` to be uploaded.
     mz_ore::test::timeout(Duration::from_secs(120), || {
         println!("test_no_block: starting server");
-        let server = util::start_server(util::Config::default()).unwrap();
+        let server = test_util::start_server(test_util::Config::default()).unwrap();
 
         server.runtime.block_on(async {
             println!("test_no_block: starting mock HTTP server");
@@ -245,7 +242,7 @@ fn test_no_block() {
 /// does not crash the server.
 #[mz_ore::test]
 fn test_drop_connection_race() {
-    let server = util::start_server(util::Config::default().unsafe_mode()).unwrap();
+    let server = test_util::start_server(test_util::Config::default().unsafe_mode()).unwrap();
     info!("test_drop_connection_race: server started");
 
     server.runtime.block_on(async {
@@ -323,7 +320,7 @@ fn test_drop_connection_race() {
 
 #[mz_ore::test]
 fn test_time() {
-    let server = util::start_server(util::Config::default()).unwrap();
+    let server = test_util::start_server(test_util::Config::default()).unwrap();
     server.enable_feature_flags(&["enable_dangerous_functions"]);
     let mut client = server.connect(postgres::NoTls).unwrap();
 
@@ -367,8 +364,8 @@ fn test_time() {
 
 #[mz_ore::test]
 fn test_subscribe_consolidation() {
-    let config = util::Config::default().workers(2);
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default().workers(2);
+    let server = test_util::start_server(config).unwrap();
     let mut client_writes = server.connect(postgres::NoTls).unwrap();
     let mut client_reads = server.connect(postgres::NoTls).unwrap();
 
@@ -397,8 +394,8 @@ fn test_subscribe_consolidation() {
 
 #[mz_ore::test]
 fn test_subscribe_negative_diffs() {
-    let config = util::Config::default().workers(2);
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default().workers(2);
+    let server = test_util::start_server(config).unwrap();
     let mut client_writes = server.connect(postgres::NoTls).unwrap();
     let mut client_reads = server.connect(postgres::NoTls).unwrap();
 
@@ -449,8 +446,8 @@ fn test_subscribe_negative_diffs() {
 
 #[mz_ore::test]
 fn test_empty_subscribe_notice() {
-    let config = util::Config::default().with_now(NOW_ZERO.clone());
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default().with_now(NOW_ZERO.clone());
+    let server = test_util::start_server(config).unwrap();
     let (tx, mut rx) = futures::channel::mpsc::unbounded();
     let mut client = server
         .pg_config()
@@ -459,7 +456,7 @@ fn test_empty_subscribe_notice() {
         .unwrap();
 
     client.batch_execute("CREATE TABLE t (a int)").unwrap();
-    let now = util::get_explain_timestamp("t", &mut client);
+    let now = test_util::get_explain_timestamp("t", &mut client);
     client
         .batch_execute(&format!("SUBSCRIBE TO t AS OF {now} UP TO {now}"))
         .unwrap();
@@ -480,12 +477,12 @@ fn test_empty_subscribe_notice() {
 
 #[mz_ore::test]
 fn test_empty_subscribe_error() {
-    let config = util::Config::default().with_now(NOW_ZERO.clone());
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default().with_now(NOW_ZERO.clone());
+    let server = test_util::start_server(config).unwrap();
     let mut client = server.connect(postgres::NoTls).unwrap();
 
     client.batch_execute("CREATE TABLE t (a int)").unwrap();
-    let now = util::get_explain_timestamp("t", &mut client);
+    let now = test_util::get_explain_timestamp("t", &mut client);
     let e = client
         .batch_execute(&format!("SUBSCRIBE TO t AS OF {now} UP TO {}", now - 1))
         .expect_err("expected DB error");
@@ -501,11 +498,11 @@ fn test_subscribe_basic() {
         let nowfn = Arc::clone(&nowfn);
         NowFn::from(move || (nowfn.lock().unwrap())())
     };
-    let config = util::Config::default()
+    let config = test_util::Config::default()
         .workers(2)
         .with_now(now)
         .unsafe_mode();
-    let server = util::start_server(config).unwrap();
+    let server = test_util::start_server(config).unwrap();
     let mut client_writes = server.connect(postgres::NoTls).unwrap();
     let mut client_reads = server.connect(postgres::NoTls).unwrap();
 
@@ -669,8 +666,8 @@ fn test_subscribe_progress() {
     for has_initial_data in [false, true] {
         for has_index in [false, true] {
             for has_snapshot in [false, true] {
-                let config = util::Config::default().workers(2);
-                let server = util::start_server(config).unwrap();
+                let config = test_util::Config::default().workers(2);
+                let server = test_util::start_server(config).unwrap();
                 let mut client_writes = server.connect(postgres::NoTls).unwrap();
                 let mut client_reads = server.connect(postgres::NoTls).unwrap();
 
@@ -779,8 +776,8 @@ fn test_subscribe_progress() {
 // turns them into nullable columns. See #6304.
 #[mz_ore::test]
 fn test_subscribe_progress_non_nullable_columns() {
-    let config = util::Config::default().workers(2);
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default().workers(2);
+    let server = test_util::start_server(config).unwrap();
     let mut client_writes = server.connect(postgres::NoTls).unwrap();
     let mut client_reads = server.connect(postgres::NoTls).unwrap();
 
@@ -830,8 +827,8 @@ fn test_subscribe_progress_non_nullable_columns() {
 /// receive data or not.
 #[mz_ore::test]
 fn test_subcribe_continuous_progress() {
-    let config = util::Config::default().workers(2);
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default().workers(2);
+    let server = test_util::start_server(config).unwrap();
     let mut client_writes = server.connect(postgres::NoTls).unwrap();
     let mut client_reads = server.connect(postgres::NoTls).unwrap();
 
@@ -915,8 +912,8 @@ fn test_subcribe_continuous_progress() {
 
 #[mz_ore::test]
 fn test_subscribe_fetch_timeout() {
-    let config = util::Config::default().workers(2);
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default().workers(2);
+    let server = test_util::start_server(config).unwrap();
     let mut client = server.connect(postgres::NoTls).unwrap();
 
     client.batch_execute("CREATE TABLE t (i INT8)").unwrap();
@@ -1012,8 +1009,8 @@ fn test_subscribe_fetch_timeout() {
 
 #[mz_ore::test]
 fn test_subscribe_fetch_wait() {
-    let config = util::Config::default().workers(2);
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default().workers(2);
+    let server = test_util::start_server(config).unwrap();
     let mut client = server.connect(postgres::NoTls).unwrap();
 
     client.batch_execute("CREATE TABLE t (i INT8)").unwrap();
@@ -1075,8 +1072,8 @@ fn test_subscribe_fetch_wait() {
 
 #[mz_ore::test]
 fn test_subscribe_empty_upper_frontier() {
-    let config = util::Config::default();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default();
+    let server = test_util::start_server(config).unwrap();
     let mut client = server.connect(postgres::NoTls).unwrap();
 
     client
@@ -1096,7 +1093,7 @@ fn test_subscribe_empty_upper_frontier() {
 // does not keep the server alive forever.
 #[mz_ore::test]
 fn test_subscribe_shutdown() {
-    let server = util::start_server(util::Config::default()).unwrap();
+    let server = test_util::start_server(test_util::Config::default()).unwrap();
 
     // We have to use the async PostgreSQL client so that we can ungracefully
     // abort the connection task.
@@ -1136,8 +1133,8 @@ fn test_subscribe_shutdown() {
 
 #[mz_ore::test]
 fn test_subscribe_table_rw_timestamps() {
-    let config = util::Config::default().workers(3);
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default().workers(3);
+    let server = test_util::start_server(config).unwrap();
     let mut client_interactive = server.connect(postgres::NoTls).unwrap();
     let mut client_subscribe = server.connect(postgres::NoTls).unwrap();
 
@@ -1221,7 +1218,7 @@ fn test_subscribe_table_rw_timestamps() {
 // by another connection.
 #[mz_ore::test]
 fn test_temporary_views() {
-    let server = util::start_server(util::Config::default()).unwrap();
+    let server = test_util::start_server(test_util::Config::default()).unwrap();
     let mut client_a = server.connect(postgres::NoTls).unwrap();
     let mut client_b = server.connect(postgres::NoTls).unwrap();
     client_a
@@ -1251,8 +1248,8 @@ fn test_temporary_views() {
 // Test EXPLAIN TIMESTAMP with tables.
 #[mz_ore::test]
 fn test_explain_timestamp_table() {
-    let config = util::Config::default();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default();
+    let server = test_util::start_server(config).unwrap();
     let mut client = server.connect(postgres::NoTls).unwrap();
     let timestamp_re = Regex::new(r"\s*(\d+) \(\d+-\d\d-\d\d \d\d:\d\d:\d\d.\d\d\d\)").unwrap();
 
@@ -1282,8 +1279,8 @@ source materialize.public.t1 (u1, storage):
 // Test `EXPLAIN TIMESTAMP AS JSON`
 #[mz_ore::test]
 fn test_explain_timestamp_json() {
-    let config = util::Config::default();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default();
+    let server = test_util::start_server(config).unwrap();
     let mut client = server.connect(postgres::NoTls).unwrap();
     client.batch_execute("CREATE TABLE t1 (i1 int)").unwrap();
 
@@ -1305,9 +1302,9 @@ fn test_explain_timestamp_json() {
 // GitHub issue # 18950
 #[mz_ore::test]
 fn test_transactional_explain_timestamps() {
-    let config = util::Config::default().workers(2).unsafe_mode();
+    let config = test_util::Config::default().workers(2).unsafe_mode();
 
-    let server = util::start_server(config).unwrap();
+    let server = test_util::start_server(config).unwrap();
 
     let mut client_writes = server.connect(postgres::NoTls).unwrap();
     let mut client_reads = server.connect(postgres::NoTls).unwrap();
@@ -1436,12 +1433,12 @@ fn test_utilization_hold() {
         NowFn::from(move || *timestamp.lock().unwrap())
     };
     let data_dir = tempfile::tempdir().unwrap();
-    let config = util::Config::default()
+    let config = test_util::Config::default()
         .with_now(now_fn)
         .data_directory(data_dir.path());
 
     // Create the server with the past time, to make sure the table is created.
-    let server = util::start_server(config).unwrap();
+    let server = test_util::start_server(config).unwrap();
 
     // Fast-forward time to make sure the table is still readable at the old time.
     *now.lock().unwrap() = now_millis;
@@ -1522,8 +1519,8 @@ fn test_utilization_hold() {
 #[ignore] // TODO(necaris): Re-enable this as soon as possible
 #[mz_ore::test]
 fn test_github_12546() {
-    let config = util::Config::default().with_propagate_crashes(false);
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default().with_propagate_crashes(false);
+    let server = test_util::start_server(config).unwrap();
     server.enable_feature_flags(&["enable_dangerous_functions"]);
 
     server
@@ -1575,8 +1572,8 @@ fn test_github_12546() {
 
 #[mz_ore::test]
 fn test_github_12951() {
-    let config = util::Config::default();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default();
+    let server = test_util::start_server(config).unwrap();
 
     // Verify sinks (SUBSCRIBE) are correctly handled for a dropped cluster.
     {
@@ -1635,8 +1632,8 @@ fn test_github_12951() {
 #[mz_ore::test]
 // Tests github issue #13100
 fn test_subscribe_outlive_cluster() {
-    let config = util::Config::default();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default();
+    let server = test_util::start_server(config).unwrap();
 
     // Verify sinks (SUBSCRIBE) are correctly handled for a dropped cluster, when a new cluster is created.
     let mut client1 = server.connect(postgres::NoTls).unwrap();
@@ -1670,8 +1667,8 @@ fn test_subscribe_outlive_cluster() {
 
 #[mz_ore::test]
 fn test_read_then_write_serializability() {
-    let config = util::Config::default();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default();
+    let server = test_util::start_server(config).unwrap();
 
     // Create table with initial value
     {
@@ -1727,26 +1724,26 @@ fn test_timestamp_recovery() {
         NowFn::from(move || *timestamp.lock().unwrap())
     };
     let data_dir = tempfile::tempdir().unwrap();
-    let config = util::Config::default()
+    let config = test_util::Config::default()
         .with_now(now_fn)
         .data_directory(data_dir.path());
 
     // Start a server and get the current global timestamp
     let global_timestamp = {
-        let server = util::start_server(config.clone()).unwrap();
+        let server = test_util::start_server(config.clone()).unwrap();
         let mut client = server.connect(postgres::NoTls).unwrap();
 
         client.batch_execute("CREATE TABLE t1 (i1 INT)").unwrap();
-        util::get_explain_timestamp("t1", &mut client)
+        test_util::get_explain_timestamp("t1", &mut client)
     };
 
     // Rollback the current time and ensure that a value larger than the old global timestamp is
     // recovered
     {
         *now.lock().expect("lock poisoned") = 0;
-        let server = util::start_server(config).unwrap();
+        let server = test_util::start_server(config).unwrap();
         let mut client = server.connect(postgres::NoTls).unwrap();
-        let recovered_timestamp = util::get_explain_timestamp("t1", &mut client);
+        let recovered_timestamp = test_util::get_explain_timestamp("t1", &mut client);
         assert!(recovered_timestamp > global_timestamp);
     }
 }
@@ -1759,13 +1756,13 @@ fn test_timeline_read_holds() {
         let now = Arc::clone(&now);
         NowFn::from(move || *now.lock().unwrap())
     };
-    let config = util::Config::default().with_now(now_fn).unsafe_mode();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default().with_now(now_fn).unsafe_mode();
+    let server = test_util::start_server(config).unwrap();
     let mut mz_client = server.connect(postgres::NoTls).unwrap();
 
     let view_name = "v_hold";
     let source_name = "source_hold";
-    let (mut pg_client, cleanup_fn) = util::create_postgres_source_with_table(
+    let (mut pg_client, cleanup_fn) = test_util::create_postgres_source_with_table(
         &server.runtime,
         &mut mz_client,
         view_name,
@@ -1777,7 +1774,8 @@ fn test_timeline_read_holds() {
     // Create user table in Materialize.
     mz_client.batch_execute("DROP TABLE IF EXISTS t;").unwrap();
     mz_client.batch_execute("CREATE TABLE t (a INT);").unwrap();
-    util::insert_with_deterministic_timestamps("t", "(42)", &server, Arc::clone(&now)).unwrap();
+    test_util::insert_with_deterministic_timestamps("t", "(42)", &server, Arc::clone(&now))
+        .unwrap();
 
     // Insert data into source.
     let source_rows: i64 = 10;
@@ -1788,7 +1786,7 @@ fn test_timeline_read_holds() {
             .unwrap();
     }
 
-    util::wait_for_view_population(&mut mz_client, view_name, source_rows).unwrap();
+    test_util::wait_for_view_population(&mut mz_client, view_name, source_rows).unwrap();
 
     // Make sure that the table and view are joinable immediately at some timestamp.
     let mut mz_join_client = server.connect(postgres::NoTls).unwrap();
@@ -1811,13 +1809,13 @@ fn test_linearizability() {
         let now = Arc::clone(&now);
         NowFn::from(move || *now.lock().unwrap())
     };
-    let config = util::Config::default().with_now(now_fn).unsafe_mode();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default().with_now(now_fn).unsafe_mode();
+    let server = test_util::start_server(config).unwrap();
     let mut mz_client = server.connect(postgres::NoTls).unwrap();
 
     let view_name = "v_lin";
     let source_name = "source_lin";
-    let (mut pg_client, cleanup_fn) = util::create_postgres_source_with_table(
+    let (mut pg_client, cleanup_fn) = test_util::create_postgres_source_with_table(
         &server.runtime,
         &mut mz_client,
         view_name,
@@ -1831,7 +1829,7 @@ fn test_linearizability() {
         .block_on(pg_client.execute(&format!("INSERT INTO {view_name} VALUES (42);"), &[]))
         .unwrap();
 
-    util::wait_for_view_population(&mut mz_client, view_name, 1).unwrap();
+    test_util::wait_for_view_population(&mut mz_client, view_name, 1).unwrap();
 
     // The user table's write frontier will be close to zero because we use a deterministic
     // now function in this test. It may be slightly higher than zero because bootstrapping
@@ -1845,19 +1843,19 @@ fn test_linearizability() {
     mz_client
         .batch_execute("SET transaction_isolation = serializable")
         .unwrap();
-    let view_ts = util::get_explain_timestamp(view_name, &mut mz_client);
+    let view_ts = test_util::get_explain_timestamp(view_name, &mut mz_client);
     // Create user table in Materialize.
     mz_client.batch_execute("DROP TABLE IF EXISTS t;").unwrap();
     mz_client.batch_execute("CREATE TABLE t (a INT);").unwrap();
-    let join_ts = util::get_explain_timestamp(&format!("{view_name}, t"), &mut mz_client);
+    let join_ts = test_util::get_explain_timestamp(&format!("{view_name}, t"), &mut mz_client);
     // In serializable transaction isolation, read timestamps can go backwards.
     assert!(join_ts < view_ts);
 
     mz_client
         .batch_execute("SET transaction_isolation = 'strict serializable'")
         .unwrap();
-    let view_ts = util::get_explain_timestamp(view_name, &mut mz_client);
-    let join_ts = util::get_explain_timestamp(&format!("{view_name}, t"), &mut mz_client);
+    let view_ts = test_util::get_explain_timestamp(view_name, &mut mz_client);
+    let join_ts = test_util::get_explain_timestamp(&format!("{view_name}, t"), &mut mz_client);
     // Since the query on the join was done after the query on the view, it should have a higher or
     // equal timestamp in strict serializable mode.
     assert!(join_ts >= view_ts);
@@ -1865,8 +1863,8 @@ fn test_linearizability() {
     mz_client
         .batch_execute("SET transaction_isolation = serializable")
         .unwrap();
-    let view_ts = util::get_explain_timestamp(view_name, &mut mz_client);
-    let join_ts = util::get_explain_timestamp(&format!("{view_name}, t"), &mut mz_client);
+    let view_ts = test_util::get_explain_timestamp(view_name, &mut mz_client);
+    let join_ts = test_util::get_explain_timestamp(&format!("{view_name}, t"), &mut mz_client);
     // If we go back to serializable, then timestamps can revert again.
     assert!(join_ts < view_ts);
 
@@ -1875,8 +1873,8 @@ fn test_linearizability() {
 
 #[mz_ore::test]
 fn test_internal_users() {
-    let config = util::Config::default();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default();
+    let server = test_util::start_server(config).unwrap();
 
     assert!(server
         .pg_config()
@@ -1902,8 +1900,8 @@ fn test_internal_users() {
 
 #[mz_ore::test]
 fn test_internal_users_cluster() {
-    let config = util::Config::default();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default();
+    let server = test_util::start_server(config).unwrap();
 
     for (user, expected_cluster) in INTERNAL_USER_NAME_TO_DEFAULT_CLUSTER.iter() {
         let mut internal_client = server
@@ -1924,8 +1922,8 @@ fn test_internal_users_cluster() {
 // crashing
 #[mz_ore::test]
 fn test_internal_ports() {
-    let config = util::Config::default();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default();
+    let server = test_util::start_server(config).unwrap();
 
     {
         let mut external_client = server.connect(postgres::NoTls).unwrap();
@@ -1982,8 +1980,8 @@ fn test_internal_ports() {
 // needed for this test.
 #[mz_ore::test]
 fn test_alter_system_invalid_param() {
-    let config = util::Config::default();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default();
+    let server = test_util::start_server(config).unwrap();
 
     let mut mz_client = server
         .pg_config_internal()
@@ -2010,8 +2008,8 @@ fn test_alter_system_invalid_param() {
 
 #[mz_ore::test]
 fn test_concurrent_writes() {
-    let config = util::Config::default();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default();
+    let server = test_util::start_server(config).unwrap();
 
     let num_tables = 10;
 
@@ -2068,7 +2066,7 @@ fn test_concurrent_writes() {
 
 #[mz_ore::test]
 fn test_load_generator() {
-    let server = util::start_server(util::Config::default().unsafe_mode()).unwrap();
+    let server = test_util::start_server(test_util::Config::default().unsafe_mode()).unwrap();
     let mut client = server.connect(postgres::NoTls).unwrap();
 
     client
@@ -2104,8 +2102,8 @@ fn test_load_generator() {
 
 #[mz_ore::test]
 fn test_introspection_user_permissions() {
-    let config = util::Config::default();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default();
+    let server = test_util::start_server(config).unwrap();
 
     let mut external_client = server.connect(postgres::NoTls).unwrap();
     let mut introspection_client = server
@@ -2172,8 +2170,8 @@ fn test_introspection_user_permissions() {
 
 #[mz_ore::test]
 fn test_idle_in_transaction_session_timeout() {
-    let config = util::Config::default();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default();
+    let server = test_util::start_server(config).unwrap();
     server.enable_feature_flags(&["enable_dangerous_functions"]);
 
     let mut client = server.connect(postgres::NoTls).unwrap();
@@ -2262,25 +2260,25 @@ fn test_coord_startup_blocking() {
         NowFn::from(move || *timestamp.lock().expect("lock poisoned"))
     };
     let data_dir = tempfile::tempdir().unwrap();
-    let config = util::Config::default()
+    let config = test_util::Config::default()
         .with_now(now_fn)
         .data_directory(data_dir.path());
 
     // Start 3 servers and reserve the first 3 timestamp ranges.
     {
-        let server = util::start_server(config.clone()).unwrap();
+        let server = test_util::start_server(config.clone()).unwrap();
         let mut client = server.connect(postgres::NoTls).unwrap();
 
         client.query("SELECT 1", &[]).unwrap();
     };
     {
-        let server = util::start_server(config.clone()).unwrap();
+        let server = test_util::start_server(config.clone()).unwrap();
         let mut client = server.connect(postgres::NoTls).unwrap();
 
         client.query("SELECT 1", &[]).unwrap();
     };
     {
-        let server = util::start_server(config.clone()).unwrap();
+        let server = test_util::start_server(config.clone()).unwrap();
         let mut client = server.connect(postgres::NoTls).unwrap();
 
         client.query("SELECT 1", &[]).unwrap();
@@ -2289,7 +2287,7 @@ fn test_coord_startup_blocking() {
     let (tx, rx) = std::sync::mpsc::sync_channel(0);
     std::thread::spawn(move || {
         let server =
-            util::start_server(config.clone()).expect("unable to start server asynchronously");
+            test_util::start_server(config.clone()).expect("unable to start server asynchronously");
         let mut client = server
             .connect(postgres::NoTls)
             .expect("unable to start client asynchronously");
@@ -2315,8 +2313,8 @@ fn test_coord_startup_blocking() {
 
 #[mz_ore::test]
 fn test_peek_on_dropped_cluster() {
-    let config = util::Config::default().unsafe_mode();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default().unsafe_mode();
+    let server = test_util::start_server(config).unwrap();
 
     let mut read_client = server.connect(postgres::NoTls).unwrap();
     read_client.batch_execute("CREATE TABLE t ()").unwrap();
@@ -2368,8 +2366,8 @@ fn test_peek_on_dropped_cluster() {
 
 #[mz_ore::test]
 fn test_emit_timestamp_notice() {
-    let config = util::Config::default();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default();
+    let server = test_util::start_server(config).unwrap();
 
     let (tx, mut rx) = futures::channel::mpsc::unbounded();
 
@@ -2432,8 +2430,8 @@ fn test_emit_timestamp_notice() {
 
 #[mz_ore::test]
 fn test_isolation_level_notice() {
-    let config = util::Config::default();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default();
+    let server = test_util::start_server(config).unwrap();
 
     let (tx, mut rx) = futures::channel::mpsc::unbounded();
 
@@ -2469,10 +2467,10 @@ fn test_isolation_level_notice() {
 
 #[test] // allow(test-attribute)
 fn test_emit_tracing_notice() {
-    let config = util::Config::default()
+    let config = test_util::Config::default()
         .with_enable_tracing(true)
         .with_system_parameter_default("opentelemetry_filter".to_string(), "debug".to_string());
-    let server = util::start_server(config).unwrap();
+    let server = test_util::start_server(config).unwrap();
 
     let (tx, mut rx) = futures::channel::mpsc::unbounded();
 
@@ -2599,8 +2597,8 @@ fn test_subscribe_on_dropped_source() {
             .unwrap();
     }
 
-    let config = util::Config::default();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default();
+    let server = test_util::start_server(config).unwrap();
 
     test_subscribe_on_dropped_source_inner(&server, vec!["t"], "SUBSCRIBE t", |res, rx| {
         res.unwrap();
@@ -2619,8 +2617,8 @@ fn test_subscribe_on_dropped_source() {
 
 #[mz_ore::test]
 fn test_dont_drop_sinks_twice() {
-    let config = util::Config::default().workers(4);
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default().workers(4);
+    let server = test_util::start_server(config).unwrap();
 
     let (notice_tx, mut notice_rx) = futures::channel::mpsc::unbounded();
 
@@ -2677,8 +2675,8 @@ fn test_dont_drop_sinks_twice() {
 
 #[mz_ore::test]
 fn test_timelines_persist_after_failed_transaction() {
-    let config = util::Config::default().unsafe_mode();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default().unsafe_mode();
+    let server = test_util::start_server(config).unwrap();
 
     server.enable_feature_flags(&["enable_create_source_denylist_with_options"]);
 
@@ -2720,8 +2718,8 @@ fn test_timelines_persist_after_failed_transaction() {
 // we have no way to disconnect sessions using SLT.
 #[mz_ore::test]
 fn test_mz_sessions() {
-    let config = util::Config::default();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default();
+    let server = test_util::start_server(config).unwrap();
 
     let mut foo_client = server
         .pg_config()
@@ -2847,8 +2845,8 @@ fn test_mz_sessions() {
 #[mz_ore::test]
 fn test_auto_run_on_introspection_feature_enabled() {
     // unsafe_mode enables the feature as a whole
-    let config = util::Config::default().unsafe_mode();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default().unsafe_mode();
+    let server = test_util::start_server(config).unwrap();
 
     let (tx, mut rx) = futures::channel::mpsc::unbounded();
     let mut client = server
@@ -2948,8 +2946,8 @@ const INTROSPECTION_NOTICE: &str = "results from querying these objects depend o
 #[mz_ore::test]
 fn test_auto_run_on_introspection_feature_disabled() {
     // unsafe_mode enables the feature as a whole
-    let config = util::Config::default().unsafe_mode();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default().unsafe_mode();
+    let server = test_util::start_server(config).unwrap();
 
     let (tx, mut rx) = futures::channel::mpsc::unbounded();
     let mut client = server
@@ -3031,8 +3029,8 @@ fn test_auto_run_on_introspection_feature_disabled() {
 #[mz_ore::test]
 fn test_auto_run_on_introspection_per_replica_relations() {
     // unsafe_mode enables the feature as a whole
-    let config = util::Config::default().unsafe_mode();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default().unsafe_mode();
+    let server = test_util::start_server(config).unwrap();
 
     let (tx, mut rx) = futures::channel::mpsc::unbounded();
     let mut client = server
@@ -3109,8 +3107,8 @@ fn test_auto_run_on_introspection_per_replica_relations() {
 #[mz_ore::test]
 fn test_max_connections() {
     mz_ore::test::init_logging();
-    let config = util::Config::default();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default();
+    let server = test_util::start_server(config).unwrap();
 
     let mut mz_client = server
         .pg_config_internal()
@@ -3204,8 +3202,8 @@ fn test_max_connections() {
 #[mz_ore::test]
 fn test_pg_cancel_backend() {
     mz_ore::test::init_logging();
-    let config = util::Config::default();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default();
+    let server = test_util::start_server(config).unwrap();
 
     let mut mz_client = server
         .pg_config_internal()
@@ -3325,8 +3323,8 @@ fn test_pg_cancel_backend() {
 #[mz_ore::test]
 fn test_params() {
     mz_ore::test::init_logging();
-    let config = util::Config::default();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default();
+    let server = test_util::start_server(config).unwrap();
 
     let mut client = server.connect(postgres::NoTls).unwrap();
 
@@ -3380,8 +3378,8 @@ fn test_params() {
 #[mz_ore::test]
 fn test_pg_cancel_dropped_role() {
     mz_ore::test::init_logging();
-    let config = util::Config::default();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default();
+    let server = test_util::start_server(config).unwrap();
     let dropped_role = "r1";
 
     let mut query_client = server.connect(postgres::NoTls).unwrap();
@@ -3432,8 +3430,8 @@ fn test_pg_cancel_dropped_role() {
 #[mz_ore::test]
 fn test_peek_on_dropped_indexed_view() {
     mz_ore::test::init_logging();
-    let config = util::Config::default();
-    let server = util::start_server(config).unwrap();
+    let config = test_util::Config::default();
+    let server = test_util::start_server(config).unwrap();
 
     let mut ddl_client = server.connect(postgres::NoTls).unwrap();
     let mut peek_client = server.connect(postgres::NoTls).unwrap();
